Move metadata vespa tests into separate file
